Update 1-programming_terminology.md
In Part 2 of the course, in the "Debugging" section, there was a typo that I have attempted to correct. The sentence "Notice how the "sunday" in the Boolean expression has not been capitalized, but in the input it was." Should be written as the exact opposite, in this situation. The issue with the code is that the user is attempting to input the day without capitalization, but this sentence refers to it being the Boolean expression itself that wasn't capitalized.